Foober Food Ordering Website
this website is used to order food. it has a database thats stores username and password. you can order food and pay using paystack but the food wont arrive because sadly we have no staff to manage it. its just a test project and is lacking in many ways. its a group project that was developed by me and some of my course mates in school. i was incharge pf the code compilation and databease stuff. i implimented the payment while others did other stuff